delete mocha user after every test

+++ test/helper.js
@@ -9,7 +9,7 @@ const should = chai.should();
 
 const User = require('../api/models/User');
 
-const verifyListProperties = (body) => {
+const verifyListProperties = body => {
   body.should.be.a('object');
   body.should.to.have.all.keys(
     'total',
@@ -22,7 +22,7 @@ const verifyListProperties = (body) => {
   );
 };
 
-const verifyBoardProperties = (body) => {
+const verifyBoardProperties = body => {
   body.should.be.a('object');
   body.should.have.property('id');
   body.should.have.property('name');
@@ -32,7 +32,7 @@ const verifyBoardProperties = (body) => {
   body.should.have.property('tiles');
 };
 
-const verifyUserProperties = (user) => {
+const verifyUserProperties = user => {
   user.should.be.a('object');
   user.should.have.property('id');
   user.should.have.property('name');
@@ -42,16 +42,20 @@ const verifyUserProperties = (user) => {
 
 const userData = {
   name: 'cboard mocha test',
-  email: 'anything2@cboard.io',
-  password: '123456',
+  email: 'anythingUser@cboard.io',
+  password: '123456'
 };
 
-let userid = '';
+const adminData = {
+  name: 'cboard admin mocha test',
+  email: 'anythingAdmin@cboard.io',
+  password: '123456'
+};
 
 let userForgotPassword = {
   Userid: '',
   token: '',
-  password: 'newpassword',
+  password: 'newpassword'
 };
 
 const boardData = {
@@ -67,27 +71,27 @@ const boardData = {
       image: '/symbols/mulberry/correct.svg',
       id: 'HJVQMR9pX5F-',
       backgroundColor: 'rgb(255, 241, 118)',
-      label: 'yes',
+      label: 'yes'
     },
     {
       labelKey: 'symbol.descriptiveState.no',
       image: '/symbols/mulberry/no.svg',
       id: 'SkBQMRqpX5t-',
       backgroundColor: 'rgb(255, 241, 118)',
-      label: 'no',
-    },
-  ],
+      label: 'no'
+    }
+  ]
 };
 
 function prepareDb() {
   mongoose.connect('mongodb://127.0.0.1:27017/cboard-api', {
-    useNewUrlParser: true,
+    useNewUrlParser: true
   });
   const connection = mongoose.connection;
 
   return new Promise((resolve, reject) => {
-    connection.once('open', function () {
-      mongoose.connection.db.dropDatabase(function (err, result) {
+    connection.once('open', function() {
+      mongoose.connection.db.dropDatabase(function(err, result) {
         console.log('Database droped');
         resolve(true);
       });
@@ -121,7 +125,7 @@ function generateEmail() {
 async function prepareUser(server, overrides = {}) {
   const data = {
     ...userData,
-    ...overrides,
+    ...overrides
   };
 
   const createUser = await request(server)
@@ -149,7 +153,20 @@ async function prepareUser(server, overrides = {}) {
 }
 
 async function deleteMochaUser() {
-  return User.deleteOne({ email: userData.email });
+  if (await User.exists({ email: userData.email })) {
+    await User.deleteOne({ email: userData.email });
+  }
+  if (await User.exists({ email: adminData.email })) {
+    await User.deleteOne({ email: adminData.email });
+  }
+  return;
+}
+
+async function deleteMochaUserById(userid) {
+  if (await User.exists({ _id: userid })) {
+    await User.deleteOne({ _id: userid });
+  }
+  return;
 }
 
 module.exports = {
@@ -159,8 +176,10 @@ module.exports = {
   prepareDb,
   prepareUser,
   deleteMochaUser,
+  deleteMochaUserById,
   boardData,
   userData,
+  adminData,
   userForgotPassword,
-  generateEmail: generateEmail,
+  generateEmail: generateEmail
 };
